marketmind/pyproject.toml
[project]
name = "marketmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firecrawl-py>=4.6.0",
    "langchain-google-genai>=3.0.1",
    "langgraph>=1.0.2",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
